bugfix: missing include <vector>

+++ util/test/test_configuration4.cxx
@@ -10,10 +10,10 @@ int main()
 {
     {
         Configuration cfg;
-        std::vector<size_t> sizes = {2,4,8};
+        std::vector<int> sizes = {2,4,8};
         assign(cfg, sizes);
         Assert(cfg.size() == 3);
-        for (size_t ind=0; ind<sizes.size(); ++ind) {
+        for (int ind=0; ind<sizes.size(); ++ind) {
             auto c = cfg[(int)ind];
             std::cerr << ind << " " << c << "\n";
             Assert(c.asUInt64() == sizes[ind]);
@@ -21,7 +21,7 @@ int main()
     }
     {
         Configuration cfg;
-        std::map<std::string, size_t> named = {{"two",2},{"four",4},{"eight",8}};
+        std::map<std::string, int> named = {{"two",2},{"four",4},{"eight",8}};
         assign(cfg, named);
         Assert(cfg.size() == 3);
         for (const auto& [nam,siz] : named) {
@@ -32,13 +32,13 @@ int main()
     }
     {
         Configuration cfg;
-        std::vector<size_t> sizes = {2,4,8};
-        std::map<std::string, std::vector<size_t>> nameds = {{"foo", sizes}, {"bar", {1,2,3}}};
+        std::vector<int> sizes = {2,4,8};
+        std::map<std::string, std::vector<int>> nameds = {{"foo", sizes}, {"bar", {1,2,3}}};
         assign(cfg, nameds);
         Assert(cfg.size() == 2);
         for (const auto& [nam,sizs] : nameds) {
             auto cs = cfg[nam];
-            for (size_t ind=0; ind<sizes.size(); ++ind) {
+            for (int ind=0; ind<sizes.size(); ++ind) {
                 auto c = cs[(int)ind];
                 std::cerr << nam << " " << ind << " " << c << "\n";
                 Assert(c.asUInt64() == sizs[ind]);
